requirements.txt
websockets
numpy
python-dotenv
fastapi
uvicorn[standard]
jinja2
python-multipart
